# Revert "REF: remove special casing from Index.equals (always dispatchto subclass) (#35330)"