@simonjayhawkins 2020-12-18
@jreback OK to merge this as a straight revert for backport and then follow-up for master